backend/services/auth_service/requirements.txt
flask
flask_sqlalchemy
flask_bcrypt
pyjwt
psycopg2-binary
python-dotenv
requests
